move tag module to climada.util
