+++ climada/entity/__init__.py
@@ -20,7 +20,6 @@ init entity
 """
 from .exposures import *
 from .impact_funcs import *
-from .tag import *
 from .disc_rates import *
 from .measures import *
 from .entity_def import *
