+++ climada/entity/entity_def.py
@@ -25,7 +25,7 @@ import logging
 from typing import Optional
 import pandas as pd
 
-from climada.entity.tag import Tag
+from climada.util.tag import Tag
 from climada.entity.impact_funcs.impact_func_set import ImpactFuncSet
 from climada.entity.disc_rates.base import DiscRates
 from climada.entity.measures.measure_set import MeasureSet
