+++ climada/entity/exposures/test/test_base.py
@@ -30,7 +30,7 @@ from climada import CONFIG
 from climada.entity.exposures.base import Exposures, INDICATOR_IMPF, \
      INDICATOR_CENTR, add_sea, DEF_REF_YEAR, DEF_VALUE_UNIT
 from climada.entity import LitPop
-from climada.entity.tag import Tag
+from climada.util.tag import Tag
 from climada.hazard.base import Hazard, Centroids
 from climada.util.constants import ENT_TEMPLATE_XLS, ONE_LAT_KM, DEF_CRS, HAZ_DEMO_FL
 import climada.util.coordinates as u_coord
